+++ backend/src/test/java/com/digginroom/digginroom/member/domain/password/DigginRoomPasswordEncoderTest.java
@@ -1,0 +1,33 @@
+package com.digginroom.digginroom.member.domain.password;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import com.digginroom.digginroom.util.DigginRoomPasswordEncoder;
+import org.junit.jupiter.api.DisplayNameGeneration;
+import org.junit.jupiter.api.DisplayNameGenerator;
+import org.junit.jupiter.api.Test;
+
+@SuppressWarnings("NonAsciiCharacters")
+@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
+class DigginRoomPasswordEncoderTest {
+
+    DigginRoomPasswordEncoder digginRoomPasswordEncoder = new DigginRoomPasswordEncoder();
+
+    @Test
+    void 비밀번호를_암호화한다() {
+        String plainText = "파워";
+
+        String encoded = digginRoomPasswordEncoder.encode(plainText);
+
+        assertThat(encoded).hasSize(100);
+        assertThat(encoded).doesNotContain(plainText);
+    }
+
+    @Test
+    void 비밀번호가_올바른지_확인한다() {
+        String plainText = "파워";
+        String encoded = digginRoomPasswordEncoder.encode(plainText);
+
+        assertThat(digginRoomPasswordEncoder.matches(plainText, encoded)).isTrue();
+    }
+}
